Working on the inventory system

+++ The_Battle_Of_Maldon/Public/CombatAIController.h
@@ -16,7 +16,7 @@ public:
 	ACombatAIController(const FObjectInitializer& ObjectInitializer);
 	bool canAttack; 
 
-	int getWeaponDamage(ALivingEntity* targetToCheck);
+	double getWeaponDamage(ALivingEntity* targetToCheck);
 
 	float ComboButtonPressed(FString BInput, float WeaponDamage);
